Started making code nice and modular, ran into weird bug
